requirements.txt
# Lightweight requirements - KHÔNG có PyTorch/sentence-transformers
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI
google-generativeai==0.3.2

# ChromaDB (lightweight)
chromadb==0.4.18

# ONNX Runtime (CPU only - nhẹ hơn nhiều so với PyTorch)
onnxruntime

# Tokenizers & Transformers (không cài torch)
tokenizers==0.15.0
transformers==4.36.0
sentencepiece==0.1.99

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
pandas==2.1.4
openpyxl==3.1.2

# Utils
python-dotenv==1.0.0
httpx==0.25.2
numpy>=1.21.0